intermediate messy commit -- but it works!
This implements a generalisation of NMF,
where also multiplicative components are present
and applied to some of the additive components.

The way we make this work is that after every NMF step
which improves U and V, we transform the data to
log(istic) space, and apply a NMF step there.
The exponents t of the multiplicative components T
are initialised very low and update slowly, to
favor learning U and V first.
I think this slower learning of t and T was the
key to make it work.

This also changes the initialization to
   U ~ uniform()

Model of GNMF:

  X_fs ~= sum_a{ U_fa * V_as * exp(-sum_m{ T_nm * t_ms * M_am }) }

where M is a boolean mask which indicates that
t and T should be applied to additive component a.

This also allows retrieving U and V from each iteration.

This also makes the number of iteration in a jax step a parameter.